@@ -33,19 +33,27 @@
   "dependencies": {
     "baconjs": "^0.7.89",
     "cannon": "^0.6.2",
-    "chalk": "^1.1.3",
+    "chalk": "^2.4.1",
     "gulp": "^3.9.1",
     "gulp-clean": "^0.3.2",
     "gulp-typescript": "^3.1.2",
     "johnny-five": "^0.10.0",
     "json-socket": "^0.2.1",
     "socket.io-client": "^1.5.1",
-    "typescript": "^2.0.9",
+    "typescript": "^2.8.3",
     "underscore": "^1.8.3"
   },
   "optionalDependencies": {
     "nodebots-interchange": "^1.1.2",
     "raspi-io": "^6.1.0",
     "ina219": "^0.6.1"
+  },
+  "devDependencies": {
+    "@types/baconjs": "^0.7.33",
+    "@types/cannon": "^0.1.2",
+    "@types/johnny-five": "0.0.36",
+    "@types/node": "^8.0.53",
+    "@types/socket.io-client": "^1.4.32",
+    "@types/underscore": "^1.8.8"
   }
 }
