@@ -1,7 +1,6 @@
-/// <reference path="../../../typings/globals/cannon/index.d.ts" />
 import { DroneState } from "../../states/DroneState";
 import { DroneModule } from '../../interfaces/Module';
-let CANNON = require('cannon');
+import * as CANNON from "cannon";
 
 export class Simulation implements DroneModule {
     public name: string = 'Simulation';
@@ -25,7 +24,7 @@ export class Simulation implements DroneModule {
 
 	public enable(): Promise<string> {
 		return new Promise<string>((resolve, reject) => {
-            this.enabled = true; 
+            this.enabled = true;
             this.configureActions();
             this.setupSimulation();
 			resolve();
@@ -46,14 +45,14 @@ export class Simulation implements DroneModule {
         // Setup our world
         this.world = new CANNON.World();
         this.world.gravity.set(0, -9.82, 0); // m/s²
-        this.world.solver.tolerance = 0.000000001;
+        //this.world.solver.tolerance = 0.000000001;
 
-        // Create a drone shape
+        // Create a drone body with the drone shape
         this.drone = new CANNON.Body({
             mass: 5, // kg
             position: new CANNON.Vec3(0, 0.1, 0), // m
-            shape: new CANNON.Box(new CANNON.Vec3(0.5,0.25,0.5))
         });
+        this.drone.addShape(new CANNON.Box(new CANNON.Vec3(0.5,0.25,0.5)));
         this.world.addBody(this.drone);
 
         // Create engines.
@@ -66,31 +65,31 @@ export class Simulation implements DroneModule {
 
         // Ground plane
         let plane = new CANNON.Plane();
-        let groundBody = new CANNON.Body({ 
+        let groundBody = new CANNON.Body({
             mass: 0,
             position: new CANNON.Vec3(0, -0.3, 0)
         });
         groundBody.addShape(plane);
         groundBody.quaternion.setFromAxisAngle(new CANNON.Vec3(1,0,0),-Math.PI/2);
-        this.world.add(groundBody);
+        this.world.addBody(groundBody);
 
         // Start the simulation loop
         this.lastTime = Date.now();
         this.simulationLoop();
     }
-    
+
     private createEngine(x: number, y: number, z: number): CANNON.Body {
         let shape = new CANNON.Sphere(0.1);
-        let jointBody: CANNON.Body = new CANNON.Body({ 
+        let jointBody: CANNON.Body = new CANNON.Body({
             mass: 1
         });
         jointBody.addShape(shape);
         jointBody.collisionFilterGroup = 0;
         jointBody.collisionFilterMask = 0;
         jointBody.position.set(x,y,z);
-        this.world.add(jointBody);
+        this.world.addBody(jointBody);
 
-        // The cannon body constrained by the mouse joint
+        // The CANNON body constrained by the mouse joint
         let constrainedBody: CANNON.Body = this.drone;
 
         // Vector to the engine point, relative to the body
@@ -126,13 +125,13 @@ export class Simulation implements DroneModule {
         let forceFactor = 55 * airPressureFactor;
         let torqueFactor = forceFactor;
 
-        this.engines[0].applyLocalForce(new CANNON.Vec3(-enginesState.bl.throttle*torqueFactor, enginesState.bl.throttle*forceFactor, 0), localPoint);
-        this.engines[1].applyLocalForce(new CANNON.Vec3(enginesState.br.throttle*torqueFactor, enginesState.br.throttle*forceFactor, 0), localPoint);
-        this.engines[2].applyLocalForce(new CANNON.Vec3(enginesState.fr.throttle*torqueFactor, enginesState.fr.throttle*forceFactor, 0), localPoint);
-        this.engines[3].applyLocalForce(new CANNON.Vec3(-enginesState.fl.throttle*torqueFactor, enginesState.fl.throttle*forceFactor, 0), localPoint);
+        this.engines[0].applyForce(new CANNON.Vec3(-enginesState.bl.throttle*torqueFactor, enginesState.bl.throttle*forceFactor, 0), localPoint);
+        this.engines[1].applyForce(new CANNON.Vec3(enginesState.br.throttle*torqueFactor, enginesState.br.throttle*forceFactor, 0), localPoint);
+        this.engines[2].applyForce(new CANNON.Vec3(enginesState.fr.throttle*torqueFactor, enginesState.fr.throttle*forceFactor, 0), localPoint);
+        this.engines[3].applyForce(new CANNON.Vec3(-enginesState.fl.throttle*torqueFactor, enginesState.fl.throttle*forceFactor, 0), localPoint);
 
         /**
-         * Set result position and orientation. 
+         * Set result position and orientation.
          */
         this.state.simulation.position.setValue(this.drone.position);
         this.state.simulation.orientation.setValue(this.drone.quaternion);
