@@ -1,6 +1,5 @@
-/// <reference path="../../typings/globals/cannon/index.d.ts" />
 import { StateProperty } from "../helpers/StateProperty";
-let CANNON = require('cannon');
+import * as CANNON from "cannon";
 
 export class SimulationState {
     public position: StateProperty<CANNON.Vec3>;
@@ -8,6 +7,6 @@ export class SimulationState {
 
 	constructor () {
 		this.position = new StateProperty<CANNON.Vec3>(new CANNON.Vec3(0,0,0));
-		this.orientation = new StateProperty<CANNON.Quaternion>(new CANNON.Quaternion(0,0,0,1)); 
+		this.orientation = new StateProperty<CANNON.Quaternion>(new CANNON.Quaternion(0,0,0,1));
 	}
 }
